Added Functions.h and Functions.cpp, added temporary #includes to Functions.h
